add getUser to store

+++ src/pages/api/user/[id].ts
@@ -9,10 +9,18 @@ export default async function handle(
 ) {
   // check user and session
   const userId = req.query.id as string | undefined
-  if (!await userMatchesSession(userId, req, res)) return res
+  if (!userId) return res.status(400).json('Must provide user id')
+
+  // GET user
+  if (req.method === 'GET') {
+    const user = await prisma.user.findUnique({ where: { id: userId } })
+    if (!user) return res.status(400).json('User does not exist')
+    return res.status(200).json(user)
+  }
 
   // PUT user
   if (req.method === 'PUT') {
+    if (!await userMatchesSession(userId, req, res)) return res
     if (!req.body) return res.status(400).json('No update args provided')
     const data = req.body as Prisma.UserUpdateArgs
     const updatedUser = await prisma.user.update({ where: { id: userId }, data })
@@ -21,6 +29,7 @@ export default async function handle(
 
   // DELETE user
   if (req.method === 'DELETE') {
+    if (!await userMatchesSession(userId, req, res)) return res
     const deletedUser = await prisma.user.delete({ where: { id: userId } })
     return res.status(204).json(deletedUser)
   }
